+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 dd mm yyyy
 
 - New: Log character name changes.
-- Change: Character search now includes removed and renamed characters.
+- Change: Character search now includes removed and renamed characters and can find characters by ID.
 
 ## 1.18.0